requirements.txt
numpy==1.23.5
pandas==2.1.0
pydantic==1.10.12
scikit-learn==1.3.0
faiss-cpu==1.7.4
langchain==0.0.354
guidance==0.0.64
openai==0.28.0
